Dark Themes: Set the color for PaletteMid explicitly
Apparently, on macOS, the value of QPalette::Mid is set in a different
way than on the other OSes. Let's set it explicitly to ensure the same
painting of wigdets with dark themes.

The value PaletteMid=ffa0a0a0 was determined by manual trial and error,
to ensure that it does not change anything on non-macOS.

Fixes: QDS-3281
Change-Id: I7c044d3b41116e50b2c8a676f30a3b26019ad627
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>