QmlDesigner: Reload file if type information is incomplete
Unfortuatly we require this. The code model might be still parsing imports
which will lead to an error. Reacting to ModelManagerInterface::libraryInfoUpdated
and then reloading the document does fix this.
This issue has high visibility.
In case the puppet debug mode is enabled we ignore this.

Task-number: QDS-3113
Task-number: QDS-1592
Change-Id: I6cbf9d01d1702f61c6bbbb86dca546174fecc825
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>