Fix build-time location of QLiteHtml
QtCreatorAPI sets the output location relative to the project
which makes sense for external projects, but not so much for
internal sub-projects.

In this specific case it cannot be built as a standalone project
at all anymore, since we use add_qtc_library but don't force
inclusion of QtCreatorAPI, so just remove the project() and related
setup.
When QLiteHtml is moved to an actual separate project and repository
this must be reworked anyhow.

Change-Id: I182de09619c2073aa3e1c9809c2251fd88b6bbd7
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>