JavaScriptFilter: Fix copying to clipboard
It seems that QStrings stored in a QVariant
can always be converted to EngineAction,
therefore this check can be removed.

The conversion then always resulted in the
value zero, which got equal to
EngineAction::Reset in commit 8d09191d2d1d.

While at it, forbid copying an empty string
to the clipboard, when the engine was aborted
due to a timeout.

Change-Id: Iaa4d9af52d4afd0e82f3b542d5f4e79bc8f6bdca
Reviewed-by: Orgad Shaneh <orgads@gmail.com>