Utils: Fix crash in environment widget
We called QTreeView::visualRect() with a potentially invalid model
index, yielding a bogus QRect, from which we derived a bogus QPos, at
which we tried to show a tooltip, which crashed Qt.
While the whole logic around that code looks dubious, for now we just fix
the crash by using a persistent model index.

Fixes: QTCREATORBUG-25170
Change-Id: I5647d0ed04bd228bb2887ef06465212545b54f57
Reviewed-by: Christian Stenger <christian.stenger@qt.io>