Suppress issues for .pro files that are excluded from exact parse
If a subdir in a subdir .pro file is excluded from the build with a
scope, it still is parsed with an exact parser
(in case cumulative parsing fails? see e.g.
440c40f53411054255f41f3efc96ec0a537e234c)

In that case we may not add issues to the issues pane from the exact
parser though, since for the subdir it is unexpected to be handled in
that situation. Post to messages pane like before (without popping it up
though).

Fixes: QTCREATORBUG-25201
Change-Id: If626a5cbc0001170f28bce89ae17394323af2a7b
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>