Add "r" as mnemonic for Build > Build for Run Configuration
Similar to f1881777b714ea6d7a8d10081668129ab4028b12, this change makes
it possible to build a specific sub-project using only the keyboard.
For CMake projects, this is a workaround for QTCREATORBUG-25231.

If that bug is ever fixed, this mnemonic still remains useful,
especially as it has a different meaning than Build Subproject: it
builds for the current run target as opposed to the sub-project
associated with the current file.

Task-number: QTCREATORBUG-25231
Change-Id: Iceb25a1c5c3242db3331c28ff0e4ca908dc04890
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>