CMake build/Linux: Fix location of separate plugins
We use GNUInstallDirs to define the location of plugins, libraries and
so on. If Qt Creator was configured with a specific set of locations,
we need to pass that on to separately built plugins even if
GNUInstallDirs would use different locations for them.
Otherwise Qt Creator cannot find these plugins.

Amends c81baf1a9cc938a283f6c52c8fd10bab84183391

Fixes: QTCREATORBUG-25266
Change-Id: Ib0da1428ec676fd54cb264b1aba4791b18b2c3bd
Reviewed-by: Cristian Adam <cristian.adam@qt.io>