Fix synchronization mechanism in DirectoryFilter
Currently m_directories, m_filters and m_exclusionFilters
are being modified only in the main thread, while in the
other thread they are only read. The access to them is
being protected with mutex when reading in the other thread.
However, we don't lock the mutex when modifying them in
the main thread, so it may happen that the other thread locks
the mutex and reads the value while the main thread directly
modifies it without a lock.

As a remedy, we add lacking lock mutex when we modify it from
the main thread.

In case of m_files its value is being modified in the other
thread, so we add a protection when we use it in the main thread.
In order to avoid a deadlock we unlock the mutex by the end of
restoreState(), before calling updateFileIterator() which now
relocks the mutex.

This patch may potentially fix a crash reported in
QTCREATORBUG-25265 - the last two stack traces from two different
threads, having both a call to Core::DirectoryFilter::refresh().

Task-number: QTCREATORBUG-25265
Change-Id: Ic1adfb71c2325ae06e6c2303d41b758df31d012a
Reviewed-by: Eike Ziller <eike.ziller@qt.io>