FileSystemFilter: store the value of includeHidden in prepareSearch()
Since we don't guarantee the synchronization between a call
to openConfigDialog() and matchesFor(), it may happen that
these two methods are called simultaneously by two different
threads. In theory, the value of m_includeHidden may be changed
in gui thread by a call to openConfigDialog(), while at the same
time it may be read in other thread, inside matchesFor().
As a safe solution we make a copy of this value inside prepareSearch()
and use the copy later in matchesFor().

Change-Id: I83fb184ff1914e4c02ee799464ff0d89350ab49c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>