Prevent locator widgets from running in parallel
We make a promise to locator filters that their prepareSearch() and
matchesFor() methods are never run in parallel, and never more than a
single matchesFor() call is running at the same time.

This broke with the introduction of multiple locator widgets (in
multiple windows).

Extend the existing code that makes a locator widget "wait" for a
running search to cancel and finish, to wait for any locator widget to
cancel and finish. Save the currently running search in shared future
for that purpose.

Change-Id: Id81f31faad9321cbd42bd451c452e542c99bde3a
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>