Output panes: Fix auto-scrolling issue
The idea is that the user turns auto-scrolling off by manually moving the
scrollbar up and turns it back on again by moving it all the way down.
However, at least on my machine I am no longer able to do the
latter -- the scrollbar position is always at least one below the
maximum. I don't know when, why or how that changed, nor do I know
whether it's about Qt, the window system or something else. Either way,
we have to account for it.

Change-Id: I83aab758d72735d4db45d228eca8bbc7a8eebd93
Reviewed-by: Christian Stenger <christian.stenger@qt.io>