EditorManager: Do not switch mode in response to editor closing
Switching to design mode when in process of closing editor can result
in state where there is no active text editor for a document, which
is required to properly initialize design mode. This happens when
the next editor to be activated targets the same document as the
closed editor, i.e. when the last open document is used in split view.

To prevent design mode initialization crash, we block
changing to design mode in response to editor closing, if the
new editor targets the same document as the closed editor.

Fixes: QDS-3763
Change-Id: I2bd325b805d3e526778f8cdd8e583e87b8c93a24
Reviewed-by: Eike Ziller <eike.ziller@qt.io>