PluginManager: Remove plugin crash check at shutdown
The check was mostly introduced for the case where Qt Creator crashes at
startup. For example if the user enabled or installed a plugin that
makes Qt Creator crash at startup, they get the chance to temporarily
disable that plugin because otherwise they cannot even access the plugin
dialog.

For shutdown that makes less sense, and is irritating for development.

Change-Id: I7267d74f4bb2d302c946a7488cc645ca4c7f864b
Reviewed-by: David Schulz <david.schulz@qt.io>